Restructured how we know which directories we're doing thing in to be managed by the Makefile.

+++ dl.sh
@@ -5,5 +5,10 @@ if [ "$1" == "" ]; then
 	exit 1
 fi
 
+if [ -f $1 ]; then
+	echo 'error: file exists'
+	exit 1
+fi
+
 url='http://mirror.us.leaseweb.net/ubuntu-cdimage/xubuntu/releases/17.10/release/xubuntu-17.10.1-desktop-amd64.iso'
 curl -# -o $1 $url
